// micro_services/micro_service_apis_campaign/src/domain/entities/interfaces/email-campaign/filters-to-save.interface.ts
export default interface FiltersToSave {
    activeFiltersKey: ActiveFilterKey[];
    activeFiltersValues: ActiveFiltersValues;
}

export type ActiveFilterKey = | 'operator' | 'plan' | 'contractStatus' | 'uf' | 'ageRange' | 'validity';

export type ActiveFiltersValues = Partial<{
    operator: number[];
    plan: number[];
    contractStatus: number[];
    uf: number[];
    ageRange: number[];
    validity: string[];
}>;